Add a new `DefinePropertyOverride` member to ProvidedTypeDefinition
This matches the existing `DefineMethodOverride` and its use for
implementing interface members. This additionally required marking
getter and setter methods as virtual within the `ProvidedProperty` type
definition.

Note that when the two property values specified have differing types,
an exception is thrown. This prevents e.g. using an `int` property to
override a `string` property.